requirements.txt
flask
gunicorn
google-genai
requests 